the product is a long

+++ problem-008/src/main/java/com/fde/Product.java
@@ -18,8 +18,8 @@ public class Product {
                 '}';
     }
 
-    public int getTotal() {
-        int total = 1;
+    public long getTotal() {
+        long total = 1;
 
         for (Integer factor : factors) {
             total *= factor;
